+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:bimarresan/pages/signIn.dart';
+import 'package:bimarresan/pages/welcomeFile.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -20,7 +21,7 @@ class MyApp extends StatelessWidget {
       ),
       debugShowCheckedModeBanner: false,
       title: 'bimar-resan',
-      home: SignInPage(),
+      home: WelcomePage(),
     );
   }
 }
